+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added DeviceTrigger entity
 - fix bug in parsing of device options
 
 [0.1.0] - 2025-11-30